+++ packages/superset-ui-plugins-demo/.storybook/addons.js
@@ -1,4 +1,5 @@
 // note that the import order here determines the order in the UI!
+import '@storybook/addon-actions/register';
+import '@storybook/addon-links/register';
 import '@storybook/addon-knobs/register';
-import '@storybook/addon-options/register';
 import 'storybook-addon-jsx/register';
