+++ packages/superset-ui-plugins-demo/.storybook/config.js
@@ -1,14 +1,12 @@
-import { addDecorator } from '@storybook/react';
-import { withOptions } from '@storybook/addon-options';
+import { addParameters, configure } from '@storybook/react';
 
-addDecorator(
-  withOptions({
+addParameters({
+  options: {
+    name: '@superset-ui/plugins 🔌💡',
     addonPanelInRight: false,
     enableShortcuts: false,
     goFullScreen: false,
     hierarchySeparator: /\|/,
-    // hierarchyRootSeparator: null,
-    name: '@superset-ui/plugins 🔌💡',
     selectedAddonPanel: undefined, // The order of addons in the "Addon panel" is the same as you import them in 'addons.js'. The first panel will be opened by default as you run Storybook
     showAddonPanel: true,
     showSearchBox: false,
@@ -16,5 +14,12 @@ addDecorator(
     sidebarAnimations: true,
     sortStoriesByKind: false,
     url: '#',
-  }),
-);
+  },
+});
+
+function loadStorybook() {
+  require('./storybook.css');
+  require('../storybook/stories'); // all of the stories
+}
+
+configure(loadStorybook, module);
